requirements.txt
streamlit==1.51.0
pandas==2.3.3
joblib==1.5.2
scikit-learn==1.6.1
google-generativeai==0.8.5
numpy==2.3.4
pillow==12.0.0
protobuf==5.29.5
requests==2.32.5
tqdm==4.67.1
typing-extensions==4.15.0
google-api-core==2.28.1
google-auth==2.43.0
google-api-python-client==2.187.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
pydantic==2.12.4
python-dateutil==2.9.0
pytz==2025.2
openai==1.66.3
python-dotenv==1.2.1